Rename core_map_representation_interface --> map_representation_interface
