--- ros_ws/src/robot/autonomy/planning/local/core_map_representation_interface/src/core_map_representation_interface.cpp
@@ -1,3 +1,0 @@
-#include <core_map_representation_interface/map_representation.h>
-
-int blank() { return 0; }
