+++ ros_ws/src/robot/autonomy/planning/local/disparity_map_representation/disparity_map_representation_plugin.xml
@@ -1,5 +1,6 @@
 <library path="disparity_map_representation">
-  <class type="disparity_map_representation::DisparityMapRepresentation" base_class_type="core_map_representation_interface::MapRepresentation">
+  <class type="disparity_map_representation::DisparityMapRepresentation"
+    base_class_type="map_representation_interface::MapRepresentation">
     <description>Disparity map representation plugin.</description>
   </class>
 </library>
