+++ ros_ws/src/robot/autonomy/planning/local/disparity_map_representation/include/disparity_map_representation/disparity_map_representation.hpp
@@ -1,6 +1,6 @@
 #ifndef _DISPARITY_MAP_REPRESENTATION_
 #define _DISPARITY_MAP_REPRESENTATION_
-#include <core_map_representation_interface/map_representation.h>
+#include <map_representation_interface/map_representation.h>
 #include <message_filters/subscriber.h>
 #include <message_filters/time_synchronizer.h>
 #include <tf2/LinearMath/Transform.h>
@@ -18,9 +18,9 @@
 #include <visualization_msgs/msg/marker.hpp>
 #include <visualization_msgs/msg/marker_array.hpp>
 
-namespace disparity_map_representation{
-  class DisparityMapRepresentation : public core_map_representation_interface::MapRepresentation {
-  private:
+namespace disparity_map_representation {
+class DisparityMapRepresentation : public map_representation_interface::MapRepresentation {
+   private:
     std::unique_ptr<disparity_graph::DisparityGraph> disp_graph;
 
     visualization_msgs::msg::MarkerArray markers;
@@ -35,7 +35,7 @@ namespace disparity_map_representation{
     int obstacle_check_points;
     double obstacle_check_radius;
 
-  public:
+   public:
     DisparityMapRepresentation();
     virtual double distance_to_obstacle(geometry_msgs::msg::PoseStamped pose,
                                         tf2::Vector3 direction);
@@ -44,8 +44,8 @@ namespace disparity_map_representation{
     // virtual std::vector< std::vector<double> >
     // get_values(std::vector<airstack_msgs::TrajectoryXYZVYaw> trajectories);
     virtual std::vector<std::vector<double> > get_values(
-							 std::vector<std::vector<geometry_msgs::msg::PointStamped> > trajectories);
-  };
-}
+        std::vector<std::vector<geometry_msgs::msg::PointStamped> > trajectories);
+};
+}  // namespace disparity_map_representation
 
 #endif
