+++ ros_ws/src/robot/autonomy/planning/local/disparity_map_representation/package.xml
@@ -12,7 +12,7 @@
 
   <depend>airstack_common</depend>
   <depend>airstack_msgs</depend>
-  <depend>core_map_representation_interface</depend>
+  <depend>map_representation_interface</depend>
   <depend>cv_bridge</depend>
   <depend>disparity_graph</depend>
   <depend>geometry_msgs</depend>
