+++ ros_ws/src/robot/autonomy/planning/local/droan_local_planner/package.xml
@@ -10,7 +10,7 @@
   <buildtool_depend>ament_cmake</buildtool_depend>
 
   <depend>airstack_msgs</depend>
-  <depend>core_map_representation_interface</depend>
+  <depend>map_representation_interface</depend>
   <depend>pluginlib</depend>
   <depend>rclcpp</depend>
   <depend>std_msgs</depend>
