+++ ros_ws/src/robot/autonomy/planning/local/map_representation_interface/include/map_representation_interface/map_representation.h
@@ -6,11 +6,11 @@
 #include <rclcpp/rclcpp.hpp>
 #include <vector>
 
-namespace core_map_representation_interface {
+namespace map_representation_interface {
 
-  class MapRepresentation : public rclcpp::Node {
-  private:
-  public:
+class MapRepresentation : public rclcpp::Node {
+   private:
+   public:
     /**
        Takes in a list of trajectories and outputs a value for each waypoint in each trajectory.
        @return A vector of vectors containing values for each waypoint. There is a vector of vectors
@@ -18,12 +18,12 @@ namespace core_map_representation_interface {
     */
 
     virtual std::vector<std::vector<double> > get_values(
-							 std::vector<std::vector<geometry_msgs::msg::PointStamped> >
-							 trajectories) {  // std::vector<airstack_msgs::TrajectoryXYZVYaw> trajectories){
-      RCLCPP_ERROR(this->get_logger(), "get_values CALLED BUT NOT IMPLEMENTED");
+        std::vector<std::vector<geometry_msgs::msg::PointStamped> >
+            trajectories) {  // std::vector<airstack_msgs::TrajectoryXYZVYaw> trajectories){
+        RCLCPP_ERROR(this->get_logger(), "get_values CALLED BUT NOT IMPLEMENTED");
 
-      std::vector<std::vector<double> > values;
-      return values;
+        std::vector<std::vector<double> > values;
+        return values;
     }
 
     /**
@@ -38,11 +38,11 @@ namespace core_map_representation_interface {
 
     virtual ~MapRepresentation() {}
 
-  protected:
+   protected:
     // MapRepresentation() {}
     MapRepresentation() : Node("map_representation") {}
-  };
+};
 
-}
+}  // namespace map_representation_interface
 
 #endif
