+++ ros_ws/src/robot/autonomy/planning/local/map_representation_interface/src/map_representation_interface.cpp
@@ -1,0 +1,3 @@
+#include <map_representation_interface/map_representation.h>
+
+int blank() { return 0; }
